Move autoconfig from config to conftest
